requirements.txt

Flask==2.2.5
werkzeug==2.3.7
Flask-Cors==5.0.1
flask-mysqldb==1.0.1
deepface==0.0.93
opencv-python==4.10.0.84
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.10.0.84
numpy==1.26.4
mysql-connector-python
mediapipe==0.10.21
protobuf==4.25.3
tf-keras==2.19.0
matplotlib==3.9.2
attrs==24.2.0
jax==0.5.3
jaxlib==0.5.3
sentencepiece==0.2.0


Flask-CORS
Flask-JWT-Extended
Flask-MySQLdb
python-dotenv